Fix Line Number

+++ include/libxml/tree.h
@@ -503,7 +503,7 @@ struct _xmlNode {
     struct _xmlAttr *properties;/* properties list */
     xmlNs           *nsDef;     /* namespace definitions on this node */
     void            *psvi;	/* for type/PSVI information */
-    unsigned short   line;	/* line number */
+    unsigned long   line;	/* line number */
     unsigned short   extra;	/* extra data for XPath/XSLT */
 };
